+++ CHANGELOG.md
@@ -5,6 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.106.0] - Not released
+## Changed
+- Reduced the password reset token length to 12 bytes. Perhaps enormous tokens
+  (48 bytes = 96 hexadecimal characters) were broken in the email texts. Also
+  reduced the TTL of the token from several years :) to 8 hours.
+
 ## [1.105.0] - 2022-01-05
 ### Fixed
 - A group administrator could not delete a message with an empty body (and with